feat: compare license additions from head vs base (#9)
This adds logic that lists the added packages/licenses from the HEAD pull request compared to the base branch. The differences are then submitted as pull request comment.

This is done by generating an SBOM + fetching licenses & exporting the list as TSV (using our `grant-summarize` command). The differences are checked using [`comm`].

[`comm`]: https://www.man7.org/linux/man-pages/man1/comm.1.html